Item test junit5

+++ src/main/java/com/dambroski/webStoreProject/Itens/ItemServiceimpl.java
@@ -2,9 +2,12 @@ package com.dambroski.webStoreProject.Itens;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.dambroski.webStoreProject.error.ItemNotFoundException;
 
 @Service
 public class ItemServiceimpl implements ItemService{
@@ -16,6 +19,24 @@ public class ItemServiceimpl implements ItemService{
 	public List<Item> fetchAll() {
 		return repository.findAll();
 	}
+	
+	@Override
+	public Item getItemById(Long itemId) throws ItemNotFoundException {
+		Optional<Item> item = repository.findById(itemId);
+		if(item.isEmpty()) {
+			throw new ItemNotFoundException("Itemid " + itemId + " not found");
+		}
+		return repository.findById(itemId).get();
+	}
+
+	@Override
+	public List<Item> getItemByName(String name) throws ItemNotFoundException {
+		List<Item> list = repository.findByNameLike(name);
+		if(list.isEmpty()) {
+			throw new ItemNotFoundException("Item " + name + " not found");
+		}
+		return list;
+	}
 
 	@Override
 	public Item postItem(Item item) {
@@ -23,37 +44,31 @@ public class ItemServiceimpl implements ItemService{
 	}
 
 	@Override
-	public void deleteById(Long itemId) {
+	public void deleteById(Long itemId) throws ItemNotFoundException {
+		Optional<Item> item = repository.findById(itemId);
+		if(item.isEmpty()) {
+			throw new ItemNotFoundException("Item " + itemId + " not found");
+		}
 		repository.deleteById(itemId);
 		
 	}
 
 	@Override
-	public void updateItem(Long itemId, Item item) {
-		Item newItem = repository.findById(itemId).get();
-		
+	public void updateItem(Long itemId, Item item) throws ItemNotFoundException {
+		Optional<Item> newItem = repository.findById(itemId);
 		if(Objects.nonNull(item.getName())&& !"".equals(item.getName())) {
-			newItem.setName(item.getName());
+			newItem.get().setName(item.getName());
+			
 		}
 		
 		if(Objects.nonNull(item.getPrice())&& !"".equals(item.getPrice())) {
-			newItem.setPrice(item.getPrice());
+			newItem.get().setPrice(item.getPrice());
 		}
 		
-		repository.save(newItem);
+		repository.save(newItem.get());
 		
 	}
 
-	@Override
-	public Item getItemById(Long itemId) {
-		
-		return repository.findById(itemId).get();
-	}
 
-	@Override
-	public List<Item> getItemByName(String name) {
-		
-		return repository.findByNameLike(name);
-	}
 	
 }
